feat: improve note title display
Use first line of content as title if no H1 header is present.
This improves user experience by:
- Not requiring users to know Markdown syntax
- Using the natural first line of text as title
- Making note titles more meaningful in navigation
- Fixing issues with pinned notes showing as 'Untitled note'

Previously, notes would show as 'Untitled note' unless they had an H1
header, which was not intuitive for users. Now, any text on the first
line will be used as the title, making the behavior more natural.